config: user can pick wanted wlan name via ncurses
As a replacement for the older install.sh, `make config`
can be used to select the most pertinent network interface
name based on the output of ip.c.

In a nutshell:
* User runs `make config`
* ip.c is run
* An ncurses menu appears based on the interfaces listed
* User selects a network interface name
* A sed file is created
* Now, when netmask.zsh-theme is compiled, the selected network
  interface name is what is searched for.

Remember that the default specified in netmask.zsh-theme.tmpl searches
for interfaces that start with "wl", which are the most common.

New targets:
* config (as mentioned)
* distclean (cleans up sed files)